split up actor
