+++ python/ray/tests/test_actor.py
@@ -2,8 +2,6 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import collections
-import json
 import random
 import numpy as np
 import os
@@ -12,83 +10,13 @@ try:
     import pytest_timeout
 except ImportError:
     pytest_timeout = None
-import signal
 import sys
 import time
 
 import ray
-import ray.ray_constants as ray_constants
 import ray.test_utils
 import ray.cluster_utils
-from ray.test_utils import (relevant_errors, wait_for_condition,
-                            wait_for_errors, wait_for_pid_to_exit,
-                            run_string_as_driver, generate_internal_config_map)
-
-
-@pytest.fixture
-def ray_checkpointable_actor_cls(request):
-    checkpoint_dir = "/tmp/ray_temp_checkpoint_dir/"
-    if not os.path.isdir(checkpoint_dir):
-        os.mkdir(checkpoint_dir)
-
-    class CheckpointableActor(ray.actor.Checkpointable):
-        def __init__(self):
-            self.value = 0
-            self.resumed_from_checkpoint = False
-            self.checkpoint_dir = checkpoint_dir
-
-        def node_id(self):
-            return ray.worker.global_worker.node.unique_id
-
-        def increase(self):
-            self.value += 1
-            return self.value
-
-        def get(self):
-            return self.value
-
-        def was_resumed_from_checkpoint(self):
-            return self.resumed_from_checkpoint
-
-        def get_pid(self):
-            return os.getpid()
-
-        def should_checkpoint(self, checkpoint_context):
-            # Checkpoint the actor when value is increased to 3.
-            should_checkpoint = self.value == 3
-            return should_checkpoint
-
-        def save_checkpoint(self, actor_id, checkpoint_id):
-            actor_id, checkpoint_id = actor_id.hex(), checkpoint_id.hex()
-            # Save checkpoint into a file.
-            with open(self.checkpoint_dir + actor_id, "a+") as f:
-                print(checkpoint_id, self.value, file=f)
-
-        def load_checkpoint(self, actor_id, available_checkpoints):
-            actor_id = actor_id.hex()
-            filename = self.checkpoint_dir + actor_id
-            # Load checkpoint from the file.
-            if not os.path.isfile(filename):
-                return None
-
-            available_checkpoint_ids = [
-                c.checkpoint_id for c in available_checkpoints
-            ]
-            with open(filename, "r") as f:
-                for line in f:
-                    checkpoint_id, value = line.strip().split(" ")
-                    checkpoint_id = ray.ActorCheckpointID(
-                        ray.utils.hex_to_binary(checkpoint_id))
-                    if checkpoint_id in available_checkpoint_ids:
-                        self.value = int(value)
-                        self.resumed_from_checkpoint = True
-                        return checkpoint_id
-                return None
-
-        def checkpoint_expired(self, actor_id, checkpoint_id):
-            pass
-
-    return CheckpointableActor
+from ray.test_utils import run_string_as_driver
 
 
 def test_actor_init_error_propagated(ray_start_regular):
@@ -509,159 +437,6 @@ def test_actor_method_deletion(ray_start_regular):
     assert ray.get(Actor.remote().method.remote()) == 1
 
 
-def test_actor_deletion_with_gpus(shutdown_only):
-    ray.init(
-        num_cpus=1, num_gpus=1, object_store_memory=int(150 * 1024 * 1024))
-
-    # When an actor that uses a GPU exits, make sure that the GPU resources
-    # are released.
-
-    @ray.remote(num_gpus=1)
-    class Actor(object):
-        def getpid(self):
-            return os.getpid()
-
-    for _ in range(5):
-        # If we can successfully create an actor, that means that enough
-        # GPU resources are available.
-        a = Actor.remote()
-        ray.get(a.getpid.remote())
-
-
-def test_actor_state(ray_start_regular):
-    @ray.remote
-    class Counter(object):
-        def __init__(self):
-            self.value = 0
-
-        def increase(self):
-            self.value += 1
-
-        def value(self):
-            return self.value
-
-    c1 = Counter.remote()
-    c1.increase.remote()
-    assert ray.get(c1.value.remote()) == 1
-
-    c2 = Counter.remote()
-    c2.increase.remote()
-    c2.increase.remote()
-    assert ray.get(c2.value.remote()) == 2
-
-
-def test_actor_class_methods(ray_start_regular):
-    class Foo(object):
-        x = 2
-
-        @classmethod
-        def as_remote(cls):
-            return ray.remote(cls)
-
-        @classmethod
-        def f(cls):
-            return cls.x
-
-        @classmethod
-        def g(cls, y):
-            return cls.x + y
-
-        def echo(self, value):
-            return value
-
-    a = Foo.as_remote().remote()
-    assert ray.get(a.echo.remote(2)) == 2
-    assert ray.get(a.f.remote()) == 2
-    assert ray.get(a.g.remote(2)) == 4
-
-
-def test_resource_assignment(shutdown_only):
-    """Test to make sure that we assign resource to actors at instantiation."""
-    # This test will create 16 actors. Declaring this many CPUs initially will
-    # speed up the test because the workers will be started ahead of time.
-    ray.init(
-        num_cpus=16,
-        num_gpus=1,
-        resources={"Custom": 1},
-        object_store_memory=int(150 * 1024 * 1024))
-
-    class Actor(object):
-        def __init__(self):
-            self.resources = ray.get_resource_ids()
-
-        def get_actor_resources(self):
-            return self.resources
-
-        def get_actor_method_resources(self):
-            return ray.get_resource_ids()
-
-    decorator_resource_args = [{}, {
-        "num_cpus": 0.1
-    }, {
-        "num_gpus": 0.1
-    }, {
-        "resources": {
-            "Custom": 0.1
-        }
-    }]
-    instantiation_resource_args = [{}, {
-        "num_cpus": 0.2
-    }, {
-        "num_gpus": 0.2
-    }, {
-        "resources": {
-            "Custom": 0.2
-        }
-    }]
-    for decorator_args in decorator_resource_args:
-        for instantiation_args in instantiation_resource_args:
-            if len(decorator_args) == 0:
-                actor_class = ray.remote(Actor)
-            else:
-                actor_class = ray.remote(**decorator_args)(Actor)
-            actor = actor_class._remote(**instantiation_args)
-            actor_resources = ray.get(actor.get_actor_resources.remote())
-            actor_method_resources = ray.get(
-                actor.get_actor_method_resources.remote())
-            if len(decorator_args) == 0 and len(instantiation_args) == 0:
-                assert len(actor_resources) == 0, (
-                    "Actor should not be assigned resources.")
-                assert list(actor_method_resources.keys()) == [
-                    "CPU"
-                ], ("Actor method should only have CPUs")
-                assert actor_method_resources["CPU"][0][1] == 1, (
-                    "Actor method should default to one cpu.")
-            else:
-                if ("num_cpus" not in decorator_args
-                        and "num_cpus" not in instantiation_args):
-                    assert actor_resources["CPU"][0][1] == 1, (
-                        "Actor should default to one cpu.")
-                correct_resources = {}
-                defined_resources = decorator_args.copy()
-                defined_resources.update(instantiation_args)
-                for resource, value in defined_resources.items():
-                    if resource == "num_cpus":
-                        correct_resources["CPU"] = value
-                    elif resource == "num_gpus":
-                        correct_resources["GPU"] = value
-                    elif resource == "resources":
-                        for custom_resource, amount in value.items():
-                            correct_resources[custom_resource] = amount
-                for resource, amount in correct_resources.items():
-                    assert (actor_resources[resource][0][0] ==
-                            actor_method_resources[resource][0][0]), (
-                                "Should have assigned same {} for both actor ",
-                                "and actor method.".format(resource))
-                    assert (actor_resources[resource][0][
-                        1] == actor_method_resources[resource][0][1]), (
-                            "Should have assigned same amount of {} for both ",
-                            "actor and actor method.".format(resource))
-                    assert actor_resources[resource][0][1] == amount, (
-                        "Actor should have {amount} {resource} but has ",
-                        "{amount} {resource}".format(
-                            amount=amount, resource=resource))
-
-
 def test_multiple_actors(ray_start_regular):
     @ray.remote
     class Counter(object):
@@ -990,469 +765,6 @@ def test_actor_lifetime_load_balancing(ray_start_cluster):
 
     actors = [Actor.remote() for _ in range(num_nodes)]
     ray.get([actor.ping.remote() for actor in actors])
-
-
-@pytest.mark.skipif(
-    os.environ.get("RAY_USE_NEW_GCS") == "on",
-    reason="Failing with new GCS API on Linux.")
-def test_actor_gpus(ray_start_cluster):
-    cluster = ray_start_cluster
-    num_nodes = 3
-    num_gpus_per_raylet = 4
-    for i in range(num_nodes):
-        cluster.add_node(
-            num_cpus=10 * num_gpus_per_raylet, num_gpus=num_gpus_per_raylet)
-    ray.init(address=cluster.address)
-
-    @ray.remote(num_gpus=1)
-    class Actor1(object):
-        def __init__(self):
-            self.gpu_ids = ray.get_gpu_ids()
-
-        def get_location_and_ids(self):
-            assert ray.get_gpu_ids() == self.gpu_ids
-            return (ray.worker.global_worker.node.unique_id,
-                    tuple(self.gpu_ids))
-
-    # Create one actor per GPU.
-    actors = [Actor1.remote() for _ in range(num_nodes * num_gpus_per_raylet)]
-    # Make sure that no two actors are assigned to the same GPU.
-    locations_and_ids = ray.get(
-        [actor.get_location_and_ids.remote() for actor in actors])
-    node_names = {location for location, gpu_id in locations_and_ids}
-    assert len(node_names) == num_nodes
-    location_actor_combinations = []
-    for node_name in node_names:
-        for gpu_id in range(num_gpus_per_raylet):
-            location_actor_combinations.append((node_name, (gpu_id, )))
-    assert set(locations_and_ids) == set(location_actor_combinations)
-
-    # Creating a new actor should fail because all of the GPUs are being
-    # used.
-    a = Actor1.remote()
-    ready_ids, _ = ray.wait([a.get_location_and_ids.remote()], timeout=0.01)
-    assert ready_ids == []
-
-
-def test_actor_multiple_gpus(ray_start_cluster):
-    cluster = ray_start_cluster
-    num_nodes = 3
-    num_gpus_per_raylet = 5
-    for i in range(num_nodes):
-        cluster.add_node(
-            num_cpus=10 * num_gpus_per_raylet, num_gpus=num_gpus_per_raylet)
-    ray.init(address=cluster.address)
-
-    @ray.remote(num_gpus=2)
-    class Actor1(object):
-        def __init__(self):
-            self.gpu_ids = ray.get_gpu_ids()
-
-        def get_location_and_ids(self):
-            assert ray.get_gpu_ids() == self.gpu_ids
-            return (ray.worker.global_worker.node.unique_id,
-                    tuple(self.gpu_ids))
-
-    # Create some actors.
-    actors1 = [Actor1.remote() for _ in range(num_nodes * 2)]
-    # Make sure that no two actors are assigned to the same GPU.
-    locations_and_ids = ray.get(
-        [actor.get_location_and_ids.remote() for actor in actors1])
-    node_names = {location for location, gpu_id in locations_and_ids}
-    assert len(node_names) == num_nodes
-
-    # Keep track of which GPU IDs are being used for each location.
-    gpus_in_use = {node_name: [] for node_name in node_names}
-    for location, gpu_ids in locations_and_ids:
-        gpus_in_use[location].extend(gpu_ids)
-    for node_name in node_names:
-        assert len(set(gpus_in_use[node_name])) == 4
-
-    # Creating a new actor should fail because all of the GPUs are being
-    # used.
-    a = Actor1.remote()
-    ready_ids, _ = ray.wait([a.get_location_and_ids.remote()], timeout=0.01)
-    assert ready_ids == []
-
-    # We should be able to create more actors that use only a single GPU.
-    @ray.remote(num_gpus=1)
-    class Actor2(object):
-        def __init__(self):
-            self.gpu_ids = ray.get_gpu_ids()
-
-        def get_location_and_ids(self):
-            return (ray.worker.global_worker.node.unique_id,
-                    tuple(self.gpu_ids))
-
-    # Create some actors.
-    actors2 = [Actor2.remote() for _ in range(num_nodes)]
-    # Make sure that no two actors are assigned to the same GPU.
-    locations_and_ids = ray.get(
-        [actor.get_location_and_ids.remote() for actor in actors2])
-    names = {location for location, gpu_id in locations_and_ids}
-    assert node_names == names
-    for location, gpu_ids in locations_and_ids:
-        gpus_in_use[location].extend(gpu_ids)
-    for node_name in node_names:
-        assert len(gpus_in_use[node_name]) == 5
-        assert set(gpus_in_use[node_name]) == set(range(5))
-
-    # Creating a new actor should fail because all of the GPUs are being
-    # used.
-    a = Actor2.remote()
-    ready_ids, _ = ray.wait([a.get_location_and_ids.remote()], timeout=0.01)
-    assert ready_ids == []
-
-
-def test_actor_different_numbers_of_gpus(ray_start_cluster):
-    # Test that we can create actors on two nodes that have different
-    # numbers of GPUs.
-    cluster = ray_start_cluster
-    cluster.add_node(num_cpus=10, num_gpus=0)
-    cluster.add_node(num_cpus=10, num_gpus=5)
-    cluster.add_node(num_cpus=10, num_gpus=10)
-    ray.init(address=cluster.address)
-
-    @ray.remote(num_gpus=1)
-    class Actor1(object):
-        def __init__(self):
-            self.gpu_ids = ray.get_gpu_ids()
-
-        def get_location_and_ids(self):
-            return (ray.worker.global_worker.node.unique_id,
-                    tuple(self.gpu_ids))
-
-    # Create some actors.
-    actors = [Actor1.remote() for _ in range(0 + 5 + 10)]
-    # Make sure that no two actors are assigned to the same GPU.
-    locations_and_ids = ray.get(
-        [actor.get_location_and_ids.remote() for actor in actors])
-    node_names = {location for location, gpu_id in locations_and_ids}
-    assert len(node_names) == 2
-    for node_name in node_names:
-        node_gpu_ids = [
-            gpu_id for location, gpu_id in locations_and_ids
-            if location == node_name
-        ]
-        assert len(node_gpu_ids) in [5, 10]
-        assert set(node_gpu_ids) == {(i, ) for i in range(len(node_gpu_ids))}
-
-    # Creating a new actor should fail because all of the GPUs are being
-    # used.
-    a = Actor1.remote()
-    ready_ids, _ = ray.wait([a.get_location_and_ids.remote()], timeout=0.01)
-    assert ready_ids == []
-
-
-def test_actor_multiple_gpus_from_multiple_tasks(ray_start_cluster):
-    cluster = ray_start_cluster
-    num_nodes = 5
-    num_gpus_per_raylet = 5
-    for i in range(num_nodes):
-        cluster.add_node(
-            num_cpus=10 * num_gpus_per_raylet,
-            num_gpus=num_gpus_per_raylet,
-            _internal_config=json.dumps({
-                "num_heartbeats_timeout": 1000
-            }))
-    ray.init(address=cluster.address)
-
-    @ray.remote
-    def create_actors(i, n):
-        @ray.remote(num_gpus=1)
-        class Actor(object):
-            def __init__(self, i, j):
-                self.gpu_ids = ray.get_gpu_ids()
-
-            def get_location_and_ids(self):
-                return ((ray.worker.global_worker.node.unique_id),
-                        tuple(self.gpu_ids))
-
-            def sleep(self):
-                time.sleep(100)
-
-        # Create n actors.
-        actors = []
-        for j in range(n):
-            actors.append(Actor.remote(i, j))
-
-        locations = ray.get(
-            [actor.get_location_and_ids.remote() for actor in actors])
-
-        # Put each actor to sleep for a long time to prevent them from getting
-        # terminated.
-        for actor in actors:
-            actor.sleep.remote()
-
-        return locations
-
-    all_locations = ray.get([
-        create_actors.remote(i, num_gpus_per_raylet) for i in range(num_nodes)
-    ])
-
-    # Make sure that no two actors are assigned to the same GPU.
-    node_names = {
-        location
-        for locations in all_locations for location, gpu_id in locations
-    }
-    assert len(node_names) == num_nodes
-
-    # Keep track of which GPU IDs are being used for each location.
-    gpus_in_use = {node_name: [] for node_name in node_names}
-    for locations in all_locations:
-        for location, gpu_ids in locations:
-            gpus_in_use[location].extend(gpu_ids)
-    for node_name in node_names:
-        assert len(set(gpus_in_use[node_name])) == num_gpus_per_raylet
-
-    @ray.remote(num_gpus=1)
-    class Actor(object):
-        def __init__(self):
-            self.gpu_ids = ray.get_gpu_ids()
-
-        def get_location_and_ids(self):
-            return (ray.worker.global_worker.node.unique_id,
-                    tuple(self.gpu_ids))
-
-    # All the GPUs should be used up now.
-    a = Actor.remote()
-    ready_ids, _ = ray.wait([a.get_location_and_ids.remote()], timeout=0.01)
-    assert ready_ids == []
-
-
-@pytest.mark.skipif(
-    sys.version_info < (3, 0), reason="This test requires Python 3.")
-def test_actors_and_tasks_with_gpus(ray_start_cluster):
-    cluster = ray_start_cluster
-    num_nodes = 3
-    num_gpus_per_raylet = 2
-    for i in range(num_nodes):
-        cluster.add_node(
-            num_cpus=num_gpus_per_raylet, num_gpus=num_gpus_per_raylet)
-    ray.init(address=cluster.address)
-
-    def check_intervals_non_overlapping(list_of_intervals):
-        for i in range(len(list_of_intervals)):
-            for j in range(i):
-                first_interval = list_of_intervals[i]
-                second_interval = list_of_intervals[j]
-                # Check that list_of_intervals[i] and list_of_intervals[j]
-                # don't overlap.
-                assert first_interval[0] < first_interval[1]
-                assert second_interval[0] < second_interval[1]
-                intervals_nonoverlapping = (
-                    first_interval[1] <= second_interval[0]
-                    or second_interval[1] <= first_interval[0])
-                assert intervals_nonoverlapping, (
-                    "Intervals {} and {} are overlapping.".format(
-                        first_interval, second_interval))
-
-    @ray.remote(num_gpus=1)
-    def f1():
-        t1 = time.monotonic()
-        time.sleep(0.1)
-        t2 = time.monotonic()
-        gpu_ids = ray.get_gpu_ids()
-        assert len(gpu_ids) == 1
-        assert gpu_ids[0] in range(num_gpus_per_raylet)
-        return (ray.worker.global_worker.node.unique_id, tuple(gpu_ids),
-                [t1, t2])
-
-    @ray.remote(num_gpus=2)
-    def f2():
-        t1 = time.monotonic()
-        time.sleep(0.1)
-        t2 = time.monotonic()
-        gpu_ids = ray.get_gpu_ids()
-        assert len(gpu_ids) == 2
-        assert gpu_ids[0] in range(num_gpus_per_raylet)
-        assert gpu_ids[1] in range(num_gpus_per_raylet)
-        return (ray.worker.global_worker.node.unique_id, tuple(gpu_ids),
-                [t1, t2])
-
-    @ray.remote(num_gpus=1)
-    class Actor1(object):
-        def __init__(self):
-            self.gpu_ids = ray.get_gpu_ids()
-            assert len(self.gpu_ids) == 1
-            assert self.gpu_ids[0] in range(num_gpus_per_raylet)
-
-        def get_location_and_ids(self):
-            assert ray.get_gpu_ids() == self.gpu_ids
-            return (ray.worker.global_worker.node.unique_id,
-                    tuple(self.gpu_ids))
-
-    def locations_to_intervals_for_many_tasks():
-        # Launch a bunch of GPU tasks.
-        locations_ids_and_intervals = ray.get(
-            [f1.remote() for _ in range(5 * num_nodes * num_gpus_per_raylet)] +
-            [f2.remote() for _ in range(5 * num_nodes * num_gpus_per_raylet)] +
-            [f1.remote() for _ in range(5 * num_nodes * num_gpus_per_raylet)])
-
-        locations_to_intervals = collections.defaultdict(lambda: [])
-        for location, gpu_ids, interval in locations_ids_and_intervals:
-            for gpu_id in gpu_ids:
-                locations_to_intervals[(location, gpu_id)].append(interval)
-        return locations_to_intervals
-
-    # Run a bunch of GPU tasks.
-    locations_to_intervals = locations_to_intervals_for_many_tasks()
-    # For each GPU, verify that the set of tasks that used this specific
-    # GPU did not overlap in time.
-    for locations in locations_to_intervals:
-        check_intervals_non_overlapping(locations_to_intervals[locations])
-
-    # Create an actor that uses a GPU.
-    a = Actor1.remote()
-    actor_location = ray.get(a.get_location_and_ids.remote())
-    actor_location = (actor_location[0], actor_location[1][0])
-    # This check makes sure that actor_location is formatted the same way
-    # that the keys of locations_to_intervals are formatted.
-    assert actor_location in locations_to_intervals
-
-    # Run a bunch of GPU tasks.
-    locations_to_intervals = locations_to_intervals_for_many_tasks()
-    # For each GPU, verify that the set of tasks that used this specific
-    # GPU did not overlap in time.
-    for locations in locations_to_intervals:
-        check_intervals_non_overlapping(locations_to_intervals[locations])
-    # Make sure that the actor's GPU was not used.
-    assert actor_location not in locations_to_intervals
-
-    # Create more actors to fill up all the GPUs.
-    more_actors = [
-        Actor1.remote() for _ in range(num_nodes * num_gpus_per_raylet - 1)
-    ]
-    # Wait for the actors to finish being created.
-    ray.get([actor.get_location_and_ids.remote() for actor in more_actors])
-
-    # Now if we run some GPU tasks, they should not be scheduled.
-    results = [f1.remote() for _ in range(30)]
-    ready_ids, remaining_ids = ray.wait(results, timeout=1.0)
-    assert len(ready_ids) == 0
-
-
-def test_actors_and_tasks_with_gpus_version_two(shutdown_only):
-    # Create tasks and actors that both use GPUs and make sure that they
-    # are given different GPUs
-    num_gpus = 4
-
-    ray.init(
-        num_cpus=(num_gpus + 1),
-        num_gpus=num_gpus,
-        object_store_memory=int(150 * 1024 * 1024))
-
-    # The point of this actor is to record which GPU IDs have been seen. We
-    # can't just return them from the tasks, because the tasks don't return
-    # for a long time in order to make sure the GPU is not released
-    # prematurely.
-    @ray.remote
-    class RecordGPUs(object):
-        def __init__(self):
-            self.gpu_ids_seen = []
-            self.num_calls = 0
-
-        def add_ids(self, gpu_ids):
-            self.gpu_ids_seen += gpu_ids
-            self.num_calls += 1
-
-        def get_gpu_ids_and_calls(self):
-            return self.gpu_ids_seen, self.num_calls
-
-    @ray.remote(num_gpus=1)
-    def f(record_gpu_actor):
-        gpu_ids = ray.get_gpu_ids()
-        assert len(gpu_ids) == 1
-        record_gpu_actor.add_ids.remote(gpu_ids)
-        # Sleep for a long time so that the GPU never gets released. This task
-        # will be killed by ray.shutdown() before it actually finishes.
-        time.sleep(1000)
-
-    @ray.remote(num_gpus=1)
-    class Actor(object):
-        def __init__(self, record_gpu_actor):
-            self.gpu_ids = ray.get_gpu_ids()
-            assert len(self.gpu_ids) == 1
-            record_gpu_actor.add_ids.remote(self.gpu_ids)
-
-        def check_gpu_ids(self):
-            assert ray.get_gpu_ids() == self.gpu_ids
-
-    record_gpu_actor = RecordGPUs.remote()
-
-    actors = []
-    actor_results = []
-    for _ in range(num_gpus // 2):
-        f.remote(record_gpu_actor)
-        a = Actor.remote(record_gpu_actor)
-        actor_results.append(a.check_gpu_ids.remote())
-        # Prevent the actor handle from going out of scope so that its GPU
-        # resources don't get released.
-        actors.append(a)
-
-    # Make sure that the actor method calls succeeded.
-    ray.get(actor_results)
-
-    start_time = time.time()
-    while time.time() - start_time < 30:
-        seen_gpu_ids, num_calls = ray.get(
-            record_gpu_actor.get_gpu_ids_and_calls.remote())
-        if num_calls == num_gpus:
-            break
-    assert set(seen_gpu_ids) == set(range(num_gpus))
-
-
-def test_blocking_actor_task(shutdown_only):
-    ray.init(
-        num_cpus=1, num_gpus=1, object_store_memory=int(150 * 1024 * 1024))
-
-    @ray.remote(num_gpus=1)
-    def f():
-        return 1
-
-    @ray.remote
-    class Foo(object):
-        def __init__(self):
-            pass
-
-        def blocking_method(self):
-            ray.get(f.remote())
-
-    # Make sure we can execute a blocking actor method even if there is
-    # only one CPU.
-    actor = Foo.remote()
-    ray.get(actor.blocking_method.remote())
-
-    @ray.remote(num_cpus=1)
-    class CPUFoo(object):
-        def __init__(self):
-            pass
-
-        def blocking_method(self):
-            ray.get(f.remote())
-
-    # Make sure that lifetime CPU resources are not released when actors
-    # block.
-    actor = CPUFoo.remote()
-    x_id = actor.blocking_method.remote()
-    ready_ids, remaining_ids = ray.wait([x_id], timeout=1.0)
-    assert ready_ids == []
-    assert remaining_ids == [x_id]
-
-    @ray.remote(num_gpus=1)
-    class GPUFoo(object):
-        def __init__(self):
-            pass
-
-        def blocking_method(self):
-            ray.get(f.remote())
-
-    # Make sure that GPU resources are not released when actors block.
-    actor = GPUFoo.remote()
-    x_id = actor.blocking_method.remote()
-    ready_ids, remaining_ids = ray.wait([x_id], timeout=1.0)
-    assert ready_ids == []
-    assert remaining_ids == [x_id]
 
 
 def test_exception_raised_when_actor_node_dies(ray_start_cluster_head):
@@ -2078,746 +1390,6 @@ def test_register_and_get_named_actors(ray_start_regular):
     assert ray.get(f2.method.remote()) == 2
     assert ray.get(f1.method.remote()) == 3
     assert ray.get(f2.method.remote()) == 4
-
-
-@pytest.mark.skipif(
-    sys.version_info < (3, 0),
-    reason="This test is currently failing on Python 2.7.")
-def test_lifetime_and_transient_resources(ray_start_regular):
-    # This actor acquires resources only when running methods.
-    @ray.remote
-    class Actor1(object):
-        def method(self):
-            pass
-
-    # This actor acquires resources for its lifetime.
-    @ray.remote(num_cpus=1)
-    class Actor2(object):
-        def method(self):
-            pass
-
-    actor1s = [Actor1.remote() for _ in range(10)]
-    ray.get([a.method.remote() for a in actor1s])
-
-    actor2s = [Actor2.remote() for _ in range(2)]
-    results = [a.method.remote() for a in actor2s]
-    ready_ids, remaining_ids = ray.wait(
-        results, num_returns=len(results), timeout=5.0)
-    assert len(ready_ids) == 1
-
-
-def test_custom_label_placement(ray_start_cluster):
-    cluster = ray_start_cluster
-    cluster.add_node(num_cpus=2, resources={"CustomResource1": 2})
-    cluster.add_node(num_cpus=2, resources={"CustomResource2": 2})
-    ray.init(address=cluster.address)
-
-    @ray.remote(resources={"CustomResource1": 1})
-    class ResourceActor1(object):
-        def get_location(self):
-            return ray.worker.global_worker.node.unique_id
-
-    @ray.remote(resources={"CustomResource2": 1})
-    class ResourceActor2(object):
-        def get_location(self):
-            return ray.worker.global_worker.node.unique_id
-
-    node_id = ray.worker.global_worker.node.unique_id
-
-    # Create some actors.
-    actors1 = [ResourceActor1.remote() for _ in range(2)]
-    actors2 = [ResourceActor2.remote() for _ in range(2)]
-    locations1 = ray.get([a.get_location.remote() for a in actors1])
-    locations2 = ray.get([a.get_location.remote() for a in actors2])
-    for location in locations1:
-        assert location == node_id
-    for location in locations2:
-        assert location != node_id
-
-
-def test_creating_more_actors_than_resources(shutdown_only):
-    ray.init(num_cpus=10, num_gpus=2, resources={"CustomResource1": 1})
-
-    @ray.remote(num_gpus=1)
-    class ResourceActor1(object):
-        def method(self):
-            return ray.get_gpu_ids()[0]
-
-    @ray.remote(resources={"CustomResource1": 1})
-    class ResourceActor2(object):
-        def method(self):
-            pass
-
-    # Make sure the first two actors get created and the third one does
-    # not.
-    actor1 = ResourceActor1.remote()
-    result1 = actor1.method.remote()
-    ray.wait([result1])
-    actor2 = ResourceActor1.remote()
-    result2 = actor2.method.remote()
-    ray.wait([result2])
-    actor3 = ResourceActor1.remote()
-    result3 = actor3.method.remote()
-    ready_ids, _ = ray.wait([result3], timeout=0.2)
-    assert len(ready_ids) == 0
-
-    # By deleting actor1, we free up resources to create actor3.
-    del actor1
-
-    results = ray.get([result1, result2, result3])
-    assert results[0] == results[2]
-    assert set(results) == {0, 1}
-
-    # Make sure that when one actor goes out of scope a new actor is
-    # created because some resources have been freed up.
-    results = []
-    for _ in range(3):
-        actor = ResourceActor2.remote()
-        object_id = actor.method.remote()
-        results.append(object_id)
-        # Wait for the task to execute. We do this because otherwise it may
-        # be possible for the __ray_terminate__ task to execute before the
-        # method.
-        ray.wait([object_id])
-
-    ray.get(results)
-
-
-@pytest.mark.parametrize(
-    "ray_start_object_store_memory", [150 * 1024 * 1024], indirect=True)
-def test_actor_eviction(ray_start_object_store_memory):
-    object_store_memory = ray_start_object_store_memory
-
-    @ray.remote
-    class Actor(object):
-        def __init__(self):
-            pass
-
-        def create_object(self, size):
-            return np.random.rand(size)
-
-    a = Actor.remote()
-    # Submit enough methods on the actor so that they exceed the size of the
-    # object store.
-    objects = []
-    num_objects = 20
-    for _ in range(num_objects):
-        obj = a.create_object.remote(object_store_memory // num_objects)
-        objects.append(obj)
-        # Get each object once to make sure each object gets created.
-        ray.get(obj)
-
-    # Get each object again. At this point, the earlier objects should have
-    # been evicted.
-    num_evicted, num_success = 0, 0
-    for obj in objects:
-        try:
-            val = ray.get(obj)
-            assert isinstance(val, np.ndarray), val
-            num_success += 1
-        except ray.exceptions.UnreconstructableError:
-            num_evicted += 1
-    # Some objects should have been evicted, and some should still be in the
-    # object store.
-    assert num_evicted > 0
-    assert num_success > 0
-
-
-def test_actor_reconstruction(ray_start_regular):
-    """Test actor reconstruction when actor process is killed."""
-
-    @ray.remote(max_reconstructions=1)
-    class ReconstructableActor(object):
-        """An actor that will be reconstructed at most once."""
-
-        def __init__(self):
-            self.value = 0
-
-        def increase(self, delay=0):
-            time.sleep(delay)
-            self.value += 1
-            return self.value
-
-        def get_pid(self):
-            return os.getpid()
-
-    actor = ReconstructableActor.remote()
-    pid = ray.get(actor.get_pid.remote())
-    # Call increase 3 times
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-    # Call increase again with some delay.
-    result = actor.increase.remote(delay=0.5)
-    # Sleep some time to wait for the above task to start execution.
-    time.sleep(0.2)
-    # Kill actor process, while the above task is still being executed.
-    os.kill(pid, signal.SIGKILL)
-    # Check that the above task didn't fail and the actor is reconstructed.
-    assert ray.get(result) == 4
-    # Check that we can still call the actor.
-    assert ray.get(actor.increase.remote()) == 5
-    # kill actor process one more time.
-    pid = ray.get(actor.get_pid.remote())
-    os.kill(pid, signal.SIGKILL)
-    # The actor has exceeded max reconstructions, and this task should fail.
-    with pytest.raises(ray.exceptions.RayActorError):
-        ray.get(actor.increase.remote())
-
-    # Create another actor.
-    actor = ReconstructableActor.remote()
-    # Intentionlly exit the actor
-    actor.__ray_terminate__.remote()
-    # Check that the actor won't be reconstructed.
-    with pytest.raises(ray.exceptions.RayActorError):
-        ray.get(actor.increase.remote())
-
-
-def test_actor_reconstruction_without_task(ray_start_regular):
-    """Test a dead actor can be reconstructed without sending task to it."""
-
-    @ray.remote(max_reconstructions=1)
-    class ReconstructableActor(object):
-        def __init__(self, obj_ids):
-            for obj_id in obj_ids:
-                # Every time the actor gets constructed,
-                # put a new object in plasma store.
-                global_worker = ray.worker.global_worker
-                if not global_worker.core_worker.object_exists(obj_id):
-                    global_worker.put_object(1, obj_id)
-                    break
-
-        def get_pid(self):
-            return os.getpid()
-
-    obj_ids = [ray.ObjectID.from_random() for _ in range(2)]
-    actor = ReconstructableActor.remote(obj_ids)
-    # Kill the actor.
-    pid = ray.get(actor.get_pid.remote())
-    os.kill(pid, signal.SIGKILL)
-    # Wait until the actor is reconstructed.
-    assert wait_for_condition(
-        lambda: ray.worker.global_worker.core_worker.object_exists(obj_ids[1]),
-        timeout_ms=5000)
-
-
-def test_actor_reconstruction_on_node_failure(ray_start_cluster_head):
-    """Test actor reconstruction when node dies unexpectedly."""
-    cluster = ray_start_cluster_head
-    max_reconstructions = 3
-    # Add a few nodes to the cluster.
-    # Use custom resource to make sure the actor is only created on worker
-    # nodes, not on the head node.
-    for _ in range(max_reconstructions + 2):
-        cluster.add_node(
-            resources={"a": 1},
-            _internal_config=json.dumps({
-                "initial_reconstruction_timeout_milliseconds": 200,
-                "num_heartbeats_timeout": 10,
-            }),
-        )
-
-    def kill_node(node_id):
-        node_to_remove = None
-        for node in cluster.worker_nodes:
-            if node_id == node.unique_id:
-                node_to_remove = node
-        cluster.remove_node(node_to_remove)
-
-    @ray.remote(max_reconstructions=max_reconstructions, resources={"a": 1})
-    class MyActor(object):
-        def __init__(self):
-            self.value = 0
-
-        def increase(self):
-            self.value += 1
-            return self.value
-
-        def get_object_store_socket(self):
-            return ray.worker.global_worker.node.unique_id
-
-    actor = MyActor.remote()
-    # Call increase 3 times.
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-
-    for i in range(max_reconstructions):
-        object_store_socket = ray.get(actor.get_object_store_socket.remote())
-        # Kill actor's node and the actor should be reconstructed
-        # on a different node.
-        kill_node(object_store_socket)
-        # Call increase again.
-        # Check that the actor is reconstructed and value is correct.
-        assert ray.get(actor.increase.remote()) == 4 + i
-        # Check that the actor is now on a different node.
-        assert object_store_socket != ray.get(
-            actor.get_object_store_socket.remote())
-
-    # kill the node again.
-    object_store_socket = ray.get(actor.get_object_store_socket.remote())
-    kill_node(object_store_socket)
-    # The actor has exceeded max reconstructions, and this task should fail.
-    with pytest.raises(ray.exceptions.RayActorError):
-        ray.get(actor.increase.remote())
-
-
-# NOTE(hchen): we set initial_reconstruction_timeout_milliseconds to 1s for
-# this test. Because if this value is too small, suprious task reconstruction
-# may happen and cause the test fauilure. If the value is too large, this test
-# could be very slow. We can remove this once we support dynamic timeout.
-@pytest.mark.parametrize(
-    "ray_start_cluster_head", [
-        generate_internal_config_map(
-            initial_reconstruction_timeout_milliseconds=1000)
-    ],
-    indirect=True)
-def test_multiple_actor_reconstruction(ray_start_cluster_head):
-    cluster = ray_start_cluster_head
-    # This test can be made more stressful by increasing the numbers below.
-    # The total number of actors created will be
-    # num_actors_at_a_time * num_nodes.
-    num_nodes = 5
-    num_actors_at_a_time = 3
-    num_function_calls_at_a_time = 10
-
-    worker_nodes = [
-        cluster.add_node(
-            num_cpus=3,
-            _internal_config=json.dumps({
-                "initial_reconstruction_timeout_milliseconds": 200,
-                "num_heartbeats_timeout": 10,
-            })) for _ in range(num_nodes)
-    ]
-
-    @ray.remote(max_reconstructions=ray.ray_constants.INFINITE_RECONSTRUCTION)
-    class SlowCounter(object):
-        def __init__(self):
-            self.x = 0
-
-        def inc(self, duration):
-            time.sleep(duration)
-            self.x += 1
-            return self.x
-
-    # Create some initial actors.
-    actors = [SlowCounter.remote() for _ in range(num_actors_at_a_time)]
-
-    # Wait for the actors to start up.
-    time.sleep(1)
-
-    # This is a mapping from actor handles to object IDs returned by
-    # methods on that actor.
-    result_ids = collections.defaultdict(lambda: [])
-
-    # In a loop we are going to create some actors, run some methods, kill
-    # a raylet, and run some more methods.
-    for node in worker_nodes:
-        # Create some actors.
-        actors.extend(
-            [SlowCounter.remote() for _ in range(num_actors_at_a_time)])
-        # Run some methods.
-        for j in range(len(actors)):
-            actor = actors[j]
-            for _ in range(num_function_calls_at_a_time):
-                result_ids[actor].append(actor.inc.remote(j**2 * 0.000001))
-        # Kill a node.
-        cluster.remove_node(node)
-
-        # Run some more methods.
-        for j in range(len(actors)):
-            actor = actors[j]
-            for _ in range(num_function_calls_at_a_time):
-                result_ids[actor].append(actor.inc.remote(j**2 * 0.000001))
-
-    # Get the results and check that they have the correct values.
-    for _, result_id_list in result_ids.items():
-        results = list(range(1, len(result_id_list) + 1))
-        assert ray.get(result_id_list) == results
-
-
-def kill_actor(actor):
-    """A helper function that kills an actor process."""
-    pid = ray.get(actor.get_pid.remote())
-    os.kill(pid, signal.SIGKILL)
-    wait_for_pid_to_exit(pid)
-
-
-def test_checkpointing(ray_start_regular, ray_checkpointable_actor_cls):
-    """Test actor checkpointing and restoring from a checkpoint."""
-    actor = ray.remote(
-        max_reconstructions=2)(ray_checkpointable_actor_cls).remote()
-    # Call increase 3 times, triggering a checkpoint.
-    expected = 0
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-        expected += 1
-    # Assert that the actor wasn't resumed from a checkpoint.
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is False
-    # Kill actor process.
-    kill_actor(actor)
-    # Assert that the actor was resumed from a checkpoint and its value is
-    # still correct.
-    assert ray.get(actor.get.remote()) == expected
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is True
-
-    # Submit some more tasks. These should get replayed since they happen after
-    # the checkpoint.
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-        expected += 1
-    # Kill actor again and check that reconstruction still works after the
-    # actor resuming from a checkpoint.
-    kill_actor(actor)
-    assert ray.get(actor.get.remote()) == expected
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is True
-
-
-def test_remote_checkpointing(ray_start_regular, ray_checkpointable_actor_cls):
-    """Test checkpointing of a remote actor through method invocation."""
-
-    # Define a class that exposes a method to save checkpoints.
-    class RemoteCheckpointableActor(ray_checkpointable_actor_cls):
-        def __init__(self):
-            super(RemoteCheckpointableActor, self).__init__()
-            self._should_checkpoint = False
-
-        def checkpoint(self):
-            self._should_checkpoint = True
-
-        def should_checkpoint(self, checkpoint_context):
-            should_checkpoint = self._should_checkpoint
-            self._should_checkpoint = False
-            return should_checkpoint
-
-    cls = ray.remote(max_reconstructions=2)(RemoteCheckpointableActor)
-    actor = cls.remote()
-    # Call increase 3 times.
-    expected = 0
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-        expected += 1
-    # Call a checkpoint task.
-    actor.checkpoint.remote()
-    # Assert that the actor wasn't resumed from a checkpoint.
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is False
-    # Kill actor process.
-    kill_actor(actor)
-    # Assert that the actor was resumed from a checkpoint and its value is
-    # still correct.
-    assert ray.get(actor.get.remote()) == expected
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is True
-
-    # Submit some more tasks. These should get replayed since they happen after
-    # the checkpoint.
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-        expected += 1
-    # Kill actor again and check that reconstruction still works after the
-    # actor resuming from a checkpoint.
-    kill_actor(actor)
-    assert ray.get(actor.get.remote()) == expected
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is True
-
-
-def test_checkpointing_on_node_failure(ray_start_cluster_2_nodes,
-                                       ray_checkpointable_actor_cls):
-    """Test actor checkpointing on a remote node."""
-    # Place the actor on the remote node.
-    cluster = ray_start_cluster_2_nodes
-    remote_node = list(cluster.worker_nodes)
-    actor_cls = ray.remote(max_reconstructions=1)(ray_checkpointable_actor_cls)
-    actor = actor_cls.remote()
-    while (ray.get(actor.node_id.remote()) != remote_node[0].unique_id):
-        actor = actor_cls.remote()
-
-    # Call increase several times.
-    expected = 0
-    for _ in range(6):
-        ray.get(actor.increase.remote())
-        expected += 1
-    # Assert that the actor wasn't resumed from a checkpoint.
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is False
-    # Kill actor process.
-    cluster.remove_node(remote_node[0])
-    # Assert that the actor was resumed from a checkpoint and its value is
-    # still correct.
-    assert ray.get(actor.get.remote()) == expected
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is True
-
-
-def test_checkpointing_save_exception(ray_start_regular,
-                                      ray_checkpointable_actor_cls):
-    """Test actor can still be recovered if checkpoints fail to complete."""
-
-    @ray.remote(max_reconstructions=2)
-    class RemoteCheckpointableActor(ray_checkpointable_actor_cls):
-        def save_checkpoint(self, actor_id, checkpoint_context):
-            raise Exception("Intentional error saving checkpoint.")
-
-    actor = RemoteCheckpointableActor.remote()
-    # Call increase 3 times, triggering a checkpoint that will fail.
-    expected = 0
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-        expected += 1
-    # Assert that the actor wasn't resumed from a checkpoint.
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is False
-    # Kill actor process.
-    kill_actor(actor)
-    # Assert that the actor still wasn't resumed from a checkpoint and its
-    # value is still correct.
-    assert ray.get(actor.get.remote()) == expected
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is False
-
-    # Submit some more tasks. These should get replayed since they happen after
-    # the checkpoint.
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-        expected += 1
-    # Kill actor again, and check that reconstruction still works and the actor
-    # wasn't resumed from a checkpoint.
-    kill_actor(actor)
-    assert ray.get(actor.get.remote()) == expected
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is False
-
-    # Check that the checkpoint error was pushed to the driver.
-    wait_for_errors(ray_constants.CHECKPOINT_PUSH_ERROR, 1)
-
-
-def test_checkpointing_load_exception(ray_start_regular,
-                                      ray_checkpointable_actor_cls):
-    """Test actor can still be recovered if checkpoints fail to load."""
-
-    @ray.remote(max_reconstructions=2)
-    class RemoteCheckpointableActor(ray_checkpointable_actor_cls):
-        def load_checkpoint(self, actor_id, checkpoints):
-            raise Exception("Intentional error loading checkpoint.")
-
-    actor = RemoteCheckpointableActor.remote()
-    # Call increase 3 times, triggering a checkpoint that will succeed.
-    expected = 0
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-        expected += 1
-    # Assert that the actor wasn't resumed from a checkpoint because loading
-    # it failed.
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is False
-    # Kill actor process.
-    kill_actor(actor)
-    # Assert that the actor still wasn't resumed from a checkpoint and its
-    # value is still correct.
-    assert ray.get(actor.get.remote()) == expected
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is False
-
-    # Submit some more tasks. These should get replayed since they happen after
-    # the checkpoint.
-    for _ in range(3):
-        ray.get(actor.increase.remote())
-        expected += 1
-    # Kill actor again, and check that reconstruction still works and the actor
-    # wasn't resumed from a checkpoint.
-    kill_actor(actor)
-    assert ray.get(actor.get.remote()) == expected
-    assert ray.get(actor.was_resumed_from_checkpoint.remote()) is False
-
-    # Check that the checkpoint error was pushed to the driver.
-    wait_for_errors(ray_constants.CHECKPOINT_PUSH_ERROR, 1)
-
-
-@pytest.mark.parametrize(
-    "ray_start_regular",
-    # This overwrite currently isn't effective,
-    # see https://github.com/ray-project/ray/issues/3926.
-    [generate_internal_config_map(num_actor_checkpoints_to_keep=20)],
-    indirect=True,
-)
-def test_deleting_actor_checkpoint(ray_start_regular):
-    """Test deleting old actor checkpoints."""
-
-    @ray.remote
-    class CheckpointableActor(ray.actor.Checkpointable):
-        def __init__(self):
-            self.checkpoint_ids = []
-
-        def get_checkpoint_ids(self):
-            return self.checkpoint_ids
-
-        def should_checkpoint(self, checkpoint_context):
-            # Save checkpoints after every task
-            return True
-
-        def save_checkpoint(self, actor_id, checkpoint_id):
-            self.checkpoint_ids.append(checkpoint_id)
-            pass
-
-        def load_checkpoint(self, actor_id, available_checkpoints):
-            pass
-
-        def checkpoint_expired(self, actor_id, checkpoint_id):
-            assert checkpoint_id == self.checkpoint_ids[0]
-            del self.checkpoint_ids[0]
-
-    actor = CheckpointableActor.remote()
-    for i in range(19):
-        assert len(ray.get(actor.get_checkpoint_ids.remote())) == i + 1
-    for _ in range(20):
-        assert len(ray.get(actor.get_checkpoint_ids.remote())) == 20
-
-
-def test_bad_checkpointable_actor_class():
-    """Test error raised if an actor class doesn't implement all abstract
-    methods in the Checkpointable interface."""
-
-    with pytest.raises(TypeError):
-
-        @ray.remote
-        class BadCheckpointableActor(ray.actor.Checkpointable):
-            def should_checkpoint(self, checkpoint_context):
-                return True
-
-
-def test_init_exception_in_checkpointable_actor(ray_start_regular,
-                                                ray_checkpointable_actor_cls):
-    # This test is similar to test_failure.py::test_failed_actor_init.
-    # This test is used to guarantee that checkpointable actor does not
-    # break the same logic.
-    error_message1 = "actor constructor failed"
-    error_message2 = "actor method failed"
-
-    @ray.remote
-    class CheckpointableFailedActor(ray_checkpointable_actor_cls):
-        def __init__(self):
-            raise Exception(error_message1)
-
-        def fail_method(self):
-            raise Exception(error_message2)
-
-        def should_checkpoint(self, checkpoint_context):
-            return True
-
-    a = CheckpointableFailedActor.remote()
-
-    # Make sure that we get errors from a failed constructor.
-    wait_for_errors(ray_constants.TASK_PUSH_ERROR, 1)
-    errors = relevant_errors(ray_constants.TASK_PUSH_ERROR)
-    assert len(errors) == 1
-    assert error_message1 in errors[0]["message"]
-
-    # Make sure that we get errors from a failed method.
-    a.fail_method.remote()
-    wait_for_errors(ray_constants.TASK_PUSH_ERROR, 2)
-    errors = relevant_errors(ray_constants.TASK_PUSH_ERROR)
-    assert len(errors) == 2
-    assert error_message1 in errors[1]["message"]
-
-
-def test_decorated_method(ray_start_regular):
-    def method_invocation_decorator(f):
-        def new_f_invocation(args, kwargs):
-            # Split one argument into two. Return th kwargs without passing
-            # them into the actor.
-            return f([args[0], args[0]], {}), kwargs
-
-        return new_f_invocation
-
-    def method_execution_decorator(f):
-        def new_f_execution(self, b, c):
-            # Turn two arguments into one.
-            return f(self, b + c)
-
-        new_f_execution.__ray_invocation_decorator__ = (
-            method_invocation_decorator)
-        return new_f_execution
-
-    @ray.remote
-    class Actor(object):
-        @method_execution_decorator
-        def decorated_method(self, x):
-            return x + 1
-
-    a = Actor.remote()
-
-    object_id, extra = a.decorated_method.remote(3, kwarg=3)
-    assert isinstance(object_id, ray.ObjectID)
-    assert extra == {"kwarg": 3}
-    assert ray.get(object_id) == 7  # 2 * 3 + 1
-
-
-@pytest.mark.skipif(
-    pytest_timeout is None,
-    reason="Timeout package not installed; skipping test that may hang.")
-@pytest.mark.timeout(20)
-@pytest.mark.parametrize(
-    "ray_start_cluster", [{
-        "num_cpus": 1,
-        "num_nodes": 2,
-    }], indirect=True)
-def test_ray_wait_dead_actor(ray_start_cluster):
-    """Tests that methods completed by dead actors are returned as ready"""
-    cluster = ray_start_cluster
-
-    @ray.remote(num_cpus=1)
-    class Actor(object):
-        def __init__(self):
-            pass
-
-        def node_id(self):
-            return ray.worker.global_worker.node.unique_id
-
-        def ping(self):
-            time.sleep(1)
-
-    # Create some actors and wait for them to initialize.
-    num_nodes = len(cluster.list_all_nodes())
-    actors = [Actor.remote() for _ in range(num_nodes)]
-    ray.get([actor.ping.remote() for actor in actors])
-
-    # Ping the actors and make sure the tasks complete.
-    ping_ids = [actor.ping.remote() for actor in actors]
-    ray.get(ping_ids)
-    # Evict the result from the node that we're about to kill.
-    remote_node = cluster.list_all_nodes()[-1]
-    remote_ping_id = None
-    for i, actor in enumerate(actors):
-        if ray.get(actor.node_id.remote()) == remote_node.unique_id:
-            remote_ping_id = ping_ids[i]
-    ray.internal.free([remote_ping_id], local_only=True)
-    cluster.remove_node(remote_node)
-
-    # Repeatedly call ray.wait until the exception for the dead actor is
-    # received.
-    unready = ping_ids[:]
-    while unready:
-        _, unready = ray.wait(unready, timeout=0)
-        time.sleep(1)
-
-    with pytest.raises(ray.exceptions.RayActorError):
-        ray.get(ping_ids)
-
-    # Evict the result from the dead node.
-    ray.internal.free([remote_ping_id], local_only=True)
-    # Create an actor on the local node that will call ray.wait in a loop.
-    head_node_resource = "HEAD_NODE"
-    ray.experimental.set_resource(head_node_resource, 1)
-
-    @ray.remote(num_cpus=0, resources={head_node_resource: 1})
-    class ParentActor(object):
-        def __init__(self, ping_ids):
-            self.unready = ping_ids
-
-        def wait(self):
-            _, self.unready = ray.wait(self.unready, timeout=0)
-            return len(self.unready) == 0
-
-        def ping(self):
-            return
-
-    # Repeatedly call ray.wait through the local actor until the exception for
-    # the dead actor is received.
-    parent_actor = ParentActor.remote(ping_ids)
-    ray.get(parent_actor.ping.remote())
-    failure_detected = False
-    while not failure_detected:
-        failure_detected = ray.get(parent_actor.wait.remote())
 
 
 def test_detached_actor(ray_start_regular):
